project/src/requirements.txt
asttokens==3.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
cloudpickle==3.1.2
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
daemoniker==0.2.3
decorator==5.2.1
dill==0.4.0
exceptiongroup==1.3.1
executing==2.2.1
Farama-Notifications==0.0.4
filelock==3.19.1
flaky==3.8.1
fonttools==4.60.1
fsspec==2025.10.0
getch==1.0
gym==0.23.1
gym-notices==0.1.0
gymnasium==1.0.0
humanfriendly==10.0
idna==3.11
ImageHash==4.3.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.7
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.9.4
matplotlib-inline==0.2.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.23.5
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pygame==2.6.1
pyglet==2.1.11
Pygments==2.19.2
pyparsing==3.2.5
Pyro4==4.82
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.3
requests==2.32.5
scipy==1.13.1
serpent==1.42
Shimmy==2.0.0
six==1.17.0
stable_baselines3==2.4.1
stack-data==0.6.3
sympy==1.14.0
torch==2.2.2
tqdm==4.67.1
traitlets==5.14.3
typing==3.7.4.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
xmltodict==0.12.0
zipp==3.23.0
